finish off tests and seaparate out geometry stuff
