+++ src/lib/Scenes/Artwork/Components/ImageCarousel/ImageWithLoadingState.tsx
