+++ src/lib/Scenes/Artwork/Components/ImageCarousel/__tests__/ImageWithLoadingState-tests.tsx
